pyproject.toml
[tool.poetry]
name = "redactgpt"
version = "0.1.0"
description = ""
authors = ["Dexter Yan <yanshaocong@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
openai = "^0.26.4"
pyyaml = "^6.0"
spacy = "^3.5.0"
python-decouple = "^3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
